Set `.python-version` to 3.10 for pyenv users (#4456)
The contribution docs for macOS say to install Python 3.10:


https://github.com/carbon-language/carbon-lang/blob/957599b2ab036d01b325aa6e82edea0fce7b4c53/docs/project/contribution_tools.md#L118-L124

For people who are using [pyenv](https://github.com/pyenv/pyenv) instead
of Homebrew Python, this PR adds a `.python-version` file to specify
Python 3.10 for pyenv to use.

However, I also see that later in the same document, the docs only say
that Python >=3.9 is required:


https://github.com/carbon-language/carbon-lang/blob/957599b2ab036d01b325aa6e82edea0fce7b4c53/docs/project/contribution_tools.md?plain=1#L162

So in that case, feel free to just close this PR. I also saw that #778
specifically moved Carbon away from pyenv, so if pyenv is discouraged in
general, also feel free to just close this.